+++ packages/pi/test/setup.ts
@@ -1,0 +1,4 @@
+import { setRegisteredThemes } from "@tsuuanmi/pi-tui";
+import { loadBuiltinThemes } from "../src/loader/themes/index.ts";
+
+setRegisteredThemes(loadBuiltinThemes());